Prevent stale device authority reuse

+++ src/native/kernel_api/device_broker_client.zig
@@ -22,7 +22,7 @@ pub const Client = struct {
     now_ticks: u64,
     process_generation: u32,
     device_id: u64 = 0,
-    broker_generation: u32 = 0,
+    broker_generation: u64 = 0,
     next_correlation_id: u64 = 1,
 
     pub fn init(
